Fix a library dependency issue with repeat_until_timeout
